Rename an reorganise fetcher files
